fix: attempting another round of config changes for SEO
