+++ src/app/about-us/page.tsx
@@ -8,15 +8,14 @@ import CommonHero from '@/components/common/CommonHero';
 import { Hero } from '@/types/CommonHero.type';
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
-import { SEO_CONFIG, BASEURL } from '@/utils/seoConfig';
+import { SEO_CONFIG } from '@/utils/seoConfig';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
-  metadataBase: new URL(BASEURL),
   title: SEO_CONFIG.aboutUs.title,
   description: SEO_CONFIG.aboutUs.description,
   openGraph: {
-    images: [{ url: SEO_CONFIG.aboutUs.image }],
+    images: SEO_CONFIG.aboutUs.image,
   },
 };
