+++ src/app/page.tsx
@@ -11,9 +11,7 @@ import TrustedPartners from '@/components/home/TrustedPartners';
 import PassiveEngagement from '@/components/home/PassiveEngagement';
 import BeyondNumbers from '@/components/home/BeyondNumbers';
 import HeroHomepage from '@/components/home/HeroHomepage';
-import { SEO_CONFIG, BASEURL } from '@/utils/seoConfig';
 import { getNewsList } from '@/utils/ApiHelper';
-import { Metadata } from 'next';
 
 const BackToTop = dynamic(() => import('@/components/common/BackToTop'), {
   ssr: false,
@@ -24,16 +22,6 @@ const Preloader = dynamic(() => import('@/components/common/Preloader'), {
 
 const IMAGE_WIDTH_HERO_GRID = 700;
 const IMAGE_HEIGHT_HERO_GRID = 700;
-
-export const metadata: Metadata = {
-  metadataBase: new URL(BASEURL),
-  keywords: ['AthletiFi', 'Club Soccer', 'Club Football'], // TODO: update keywords
-  title: SEO_CONFIG.home.title,
-  description: SEO_CONFIG.home.description,
-  openGraph: {
-    images: [{ url: SEO_CONFIG.home.image }],
-  },
-};
 
 // Main function component for the home page
 export default async function Home() {
