+++ src/app/privacy-policy/page.tsx
@@ -2,16 +2,15 @@ import CommonHero from '@/components/common/CommonHero';
 import { Hero } from '@/types/CommonHero.type';
 import Footer from '@/components/common/Footer';
 import Header from '@/components/common/Header';
-import { SEO_CONFIG, BASEURL } from '@/utils/seoConfig';
+import { SEO_CONFIG } from '@/utils/seoConfig';
 import PrivacyPolicy from '@/components/terms/PrivacyPolicy';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
-  metadataBase: new URL(BASEURL),
   title: SEO_CONFIG.privacyPolicy.title,
   description: SEO_CONFIG.privacyPolicy.description,
   openGraph: {
-    images: [{ url: SEO_CONFIG.privacyPolicy.image }],
+    images: SEO_CONFIG.privacyPolicy.image,
   },
 };
